tests: update suites for Decimal migration + add WC-H1/M9/M1 regressions
- Money literals now bind to Decimal; constructor helpers and store-call sites
  updated to pass Decimal amounts and the new `currency:` parameter
  (addTransaction/updateTransaction/RecurringScheduleBuilder.build).
- `accuracy:` assertions on money now compare via `.doubleValue` (Decimal isn't
  FloatingPoint); exact Decimal equality used where conversion isn't involved.
- CurrencyConverter literal-arg calls pinned to the Double overload to resolve
  the new Double/Decimal overload ambiguity.
- New regressions: MoneyParser rejects inf/nan/empty and honors locale grouping
  (WC-H1/M9); calculateTotals converts per-transaction currency (WC-M1); the
  Decimal convert overload and its zero-rate guard (WC-A1).

Not yet runnable here (no Xcode); see TO_SIMO_DO.md #23.

+++ apple/WealthCompass/Tests/FinanceImportServiceTests.swift
@@ -34,7 +34,7 @@ final class FinanceImportServiceTests: XCTestCase {
     func testCommaDecimalAmountIsParsed() throws {
         // The parser swaps "," for "." (decimal comma); it does NOT strip thousands separators.
         let result = try parse(#"{"transactions":[{"type":"expense","amount":"1234,56","category":"Rent","date":"2026-06-03"}]}"#)
-        XCTAssertEqual(result.data.transactions.first?.amount ?? 0, 1234.56, accuracy: 0.001)
+        XCTAssertEqual((result.data.transactions.first?.amount ?? 0).doubleValue, 1234.56, accuracy: 0.001)
     }
 
     func testMultipleDateFormatsBothParse() throws {
